Refactor chat.py
