+++ CLAUDE.md
@@ -32,4 +32,4 @@ The frontend's `*Character` TS interfaces (`gamemAIster-frontend/src/models/Char
 
 ## Conversation memory & compression
 
-`docs/conversation_memory.html` is the onboarding doc for how a long-running session gets compressed to fit the LLM's context window (entities/scenes/summary lenses, the ledger/digest split, NPC roster cross-referencing). **Whenever you touch anything in this area — `SummaryMemory`/`SummaryChat` in `src/brain/chat.py`, the `text_entity_prompt.txt`/`text_scene_prompt_examples.txt`/`text_summary_prompt.txt`/`text_digest_prompt.txt`/`summary_provider.txt` prompt templates, the `Entity`/`Scene`/`Interaction` models in `src/brain/data_types.py`, or the corresponding persistence in `src/crud/crud.py` and `src/crud/sqlmodel.py` — update that doc in the same change.** It's meant to stay accurate enough that someone returning after a year doesn't have to re-derive the design from the code.
+`docs/conversation_memory.html` is the onboarding doc for how a long-running session gets compressed to fit the LLM's context window (entities/scenes/summary lenses, the ledger/digest split, NPC roster cross-referencing). The subsystem is split across three files by concern: `src/brain/conversation_memory.py` (`SummaryMemory`, `CompressionPrompts` — the compression engine itself), `src/brain/npc_context.py` (`get_npc_roster`/`flag_active_key_npcs` — mission-roster cross-referencing), `src/brain/chat.py` (`SummaryChat` — turn orchestration only). **Whenever you touch any of those, the `text_entity_prompt.txt`/`text_scene_prompt_examples.txt`/`text_summary_prompt.txt`/`text_digest_prompt.txt`/`summary_provider.txt` prompt templates, the `Entity`/`Scene`/`Interaction` models in `src/brain/data_types.py`, or the corresponding persistence in `src/crud/crud.py` and `src/crud/sqlmodel.py` — update that doc in the same change.** It's meant to stay accurate enough that someone returning after a year doesn't have to re-derive the design from the code.
